Multiples Figures Partial Solution & GraphController Improved
The problem was in the KruskalAnimation.py file. The error message is:UserWarning: Animation was deleted without rendering anything. This is most likely not intended. To prevent deletion, assign the Animation to a variable, e.g. `anim`, that exists until you output the Animation using `plt.show()` or `anim.save()`